requirements.txt
backoff==1.11.1
python-dateutil==2.8.2
python-gitlab==2.10.1
python-jenkins==1.7.0
tabulate==0.8.10
tqdm==4.64.1
